requirements.txt
colorthief==0.2.1
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
paho-mqtt==1.6.1

